Clean up of util.h * Breaks dependency on specific fhir versions in last remaining /cc:* non-test targets * extracts reference logic into references.h, and type macros into type_macros.h * Removes redundant functions
PiperOrigin-RevId: 317894838
