+++ cc/google/fhir/fhir_types.h
@@ -97,6 +97,9 @@ bool IsInstant(const ::google::protobuf::Descriptor* descriptor);
 bool IsUri(const ::google::protobuf::Message& message);
 bool IsUri(const ::google::protobuf::Descriptor* descriptor);
 
+bool IsPeriod(const ::google::protobuf::Message& message);
+bool IsPeriod(const ::google::protobuf::Descriptor* descriptor);
+
 bool IsQuantity(const ::google::protobuf::Message& message);
 bool IsQuantity(const ::google::protobuf::Descriptor* descriptor);
