+++ cc/google/fhir/fhir_types_test.cc
@@ -99,13 +99,14 @@ FHIR_SIMPLE_TYPE_TEST(Identifier);
 FHIR_SIMPLE_TYPE_TEST(Instant);
 FHIR_SIMPLE_TYPE_TEST(Integer);
 FHIR_SIMPLE_TYPE_TEST(PositiveInt);
-FHIR_SIMPLE_TYPE_TEST(Quantity);
-FHIR_SIMPLE_TYPE_TEST(SimpleQuantity);
 FHIR_SIMPLE_TYPE_TEST(String);
 FHIR_SIMPLE_TYPE_TEST(Time);
 FHIR_SIMPLE_TYPE_TEST(UnsignedInt);
 FHIR_SIMPLE_TYPE_TEST(Uri);
 FHIR_SIMPLE_TYPE_TEST(Xhtml);
+FHIR_SIMPLE_TYPE_TEST(Period);
+FHIR_SIMPLE_TYPE_TEST(Quantity);
+FHIR_SIMPLE_TYPE_TEST(SimpleQuantity);
 
 }  // namespace
