+++ cc/google/fhir/seqex/bundle_to_seqex_util.h
@@ -25,6 +25,7 @@
 #include "google/fhir/extensions.h"
 #include "google/fhir/status/statusor.h"
 #include "google/fhir/util.h"
+#include "proto/stu3/datatypes.pb.h"
 
 namespace google {
 namespace fhir {
@@ -36,7 +37,7 @@ template <typename ConverterTypes>
 StatusOr<std::vector<typename ConverterTypes::EventLabel>>
 ExtractLabelsFromExtensions(
     const std::set<std::string>& label_names,
-    google::protobuf::RepeatedFieldRef<stu3::proto ::Extension> extensions) {
+    google::protobuf::RepeatedFieldRef<stu3::proto::Extension> extensions) {
   std::vector<typename ConverterTypes::EventLabel> labels;
   FHIR_RETURN_IF_ERROR(google::fhir::extensions_lib::GetRepeatedFromExtension(
       extensions, &labels));
